Adding customer basket and website as new location. Next to do, adding customers as an entity then editing 2 new action for the api, basketing a product and buying a product.
What is it ?
basketing a product is a customer action to move a product to his basket, the customer will have a basket in the warehouse he take the product and can have multiple basket. (it can be a call to a create basket then a move product)

the buy product will handle an external api call for payment then if processed well it will remove (no archive now) the basket and all the product in it.